requirements.txt
PyQt6>=6.5.0
supabase>=2.0.0
matplotlib>=3.7.0
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
win10toast>=0.9
schedule>=1.2.0
cryptography>=41.0.0
PyInstaller>=6.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
